requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bidict==0.23.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
clicksend-client==5.0.78
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
frozenlist==1.6.0
GeoAlchemy2==0.17.1
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
multidict==6.4.3
mysql-connector-python==9.3.0
numpy==2.2.5
packaging==25.0
phonenumbers==9.0.4
pluggy==1.5.0
propcache==0.3.1
pyasn1==0.4.8
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.1
python-jose==3.4.0
python-multipart==0.0.20
python-socketio==5.13.0
redis==6.1.0
requests==2.32.3
rsa==4.9.1
shapely==2.1.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
twilio==9.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
wsproto==1.2.0
yarl==1.20.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
